updated event_ts

+++ src/main/scala/TVPlayEvent.scala
@@ -1,6 +1,6 @@
 case class TVPlayEvent(
 												user_id: String,
 												show_id: String,
-												event_ts: Long,
+												event_ts: Float,
 												duration_sec: Int,
 												device: String)
